requirements.txt
numpy==1.26.4
pybullet==3.2.6
customtkinter
pillow
PyOpenGL